requirements.txt
numpy
pandas
requests
pygithub
matplotlib
python-dotenv